Make tutorial commands more easily copy-pastable
Having to edit the name in a copied command before running it bugs me.
Including the output for some of the commands but not all of them
doesn't really aid understanding when all of the commands are already
commented.
The shell prompts don't help clarity with the output removed, and they
make copying a little harder.