Merge roachperf into roachprod
Merge roachperf into roachprod. Roachprod is now a one-stop shop for
managing CockroachDB clusters.

Any invocation of roachperf is now a valid invocation of roachprod, with
the following caveats:

    * roachperf commands which took a cluster name as an argument, like
      `roachperf USER-CLUSTER start`, have been reordered. The equivalent
      roachprod command is `roachprod start USER-CLUSTER`.

    * To make room for roachperf's status command, `roachprod status`
      has been renamed to `roachprod list`. The prior behavior of
      `roachprod list` is available as `roachprod list --details`.

In the interest of incremental improvement, this is the minimal change
necessary to make this merge possible. More cleanup is in order.