Prototype of nightly test infra
NB: I'm omitting the `vendor` folder from this commit because it's
too large for Github.

This is an attempt to lay out a lower layer on which to build both
nightly tests (expanded upon in some detail here) and cli-driven
tools (i.e., mostly today's roachperf). This builds, but does not
work (most methods are stubs, but a lot of the required code already
exists in `roachprod/roachperf` and can be lifted).

These will naturally exist on various levels of abstraction. Some tests
may want to mount virtual file systems (to test corruption) and invoke
adversarial I/O workloads on the machines directly. Most performance
tests are much less concerned with these details and just need to be
told a number of Postgres endpoints and machines to run load generators
on.

Please take a look at this example code and provide criticism and
feedback. My impression is that this is a clean, elegant and at the
same time malleable foundation, and every test I have considered
wanting to write (there are a lot) seemed easy to string together
with these primitives.

The goal was to have a clear separation of concerns:

1. A *library* that provisions hardware (`roachprodng`, to be renamed
    when there's less confusion about names. I would probably just call this
    `provision` but there will be an opportunity to bikeshed).
    `roachprodng` does not know anything about CockroachDB.
    Note that the meat that needs to go into `roachprod-ng` is basically
    just `roachprod`, i.e. there isn't much that will change there
    except making it a library.

    In the current code, this library exposes some details about the
    nodes (IP addresses, available mount points) and lets you SSH/scp into
    them, but not much else.

    As a convenience, `roachprodng` allows template interpolation to
    access the provisioned node state, which makes it natural to execute
    commands on the nodes. For example, `--store {{Store 0}}` expands to
    something like `--store /mnt/data0` where `/mnt/data0` could be a
    local SSD (if that's what `roachprodng` was told to set up). This
    avoids leaking provider-specific details into the tests and should
    make it relatively straightforward to target AWS later. (Aside from
    that, it also de-noises the tests and has been pleasant to work with).

2. The `binfetcher` library is used to procure binaries, which includes
`cockroach`, but also `workload`. Binaries can be put anywhere using
`roachprodng`.

3. SQL-specific operations usually rely on `cmd/workload`, which is
    invoked on the nodes to recreate and/or ingest fixtures, and serves as
    the general-purpose SQL runner, so that many tests will simply need a
    cluster and the information on where to invoke `workload`. (It's equally
    straightforward to create a manual SQL connection in a test when that
    is all that's desired: the endpoints are readily available).

4. Various other mechanisms are modular.

    For example, many tests will want to assert the continuous uptime of the
    cluster and can simply run a component that periodically verifies this
    (through SQL). Similarly, other tests may want to periodically run
    `SCRUB` (which can be done through `cmd/workload`).

    Yet another test may want to run chaos and can tap into a modular
    component that fits in easily, though they now have to make sure that
    the load they're running tolerates errors better, and they need to
    make the monitoring component aware of the chaos events.

Note that the nightly benchmark tests are intentionally a bit
simplistic.  We'll likely want to tap into the existing `testRun` struct
from `roachperf`. This is intentional to keep the noise level down; some
of that integration is likely to happen inside of `cmd/workload` anyway.