Allow template expansion for arguments to the "ssh" command
Currently the only expansion supported is `{pgurl:<nodes>}`. If
`<nodes>` is not specified, `{pgurl}` expands to the pgurls for all of
the nodes in the cluster. The expansions allow `ssh` to be used for
starting a load generator. For example:

  roachprod ssh peter-perf:4 ./kv {pgurl:1-3}

If `peter-perf` is a 3 node cluster, this expands to:

  ./kv postgres://root@10.142.0.4:26257?sslmode=disable \
       postgres://root@10.142.0.3:26257?sslmode=disable \
       postgres://root@10.142.0.2:26257?sslmode=disable

Added the `--secure` flag to the `ssh` command which affects the
expansion of `pgurl`. Might be desirable in the future to automatically
infer from the cluster itself whether it is secure or not.