Forward SIGTERM to children
This improves roachprod's scriptability from Go. Specifically, when
using os/exec.CommandContext to launch roachprod, canceling the context
will send SIGTERM to roachprod, but not to any children launched by
roachprod. This is particularly frustrating when `roachprod ssh` is
running a command that might take minutes or hours to complete.

Forward SIGTERM by creating a new process group when roachprod starts
up. Child processes will be created in this new process group, so we can
simply forward any SIGTERMs we see to the entire process group.