Reorganize roachprod into packages
This change breaks the roachprod code up into several different packages in
order to prepare for multi-cloud support.  This is a mostly-mechanical
reorganization that just moves code location and renames symbols to enable
cross-package visibility.  The only non-trivial change is to break the original
ListCloud() function into two parts to prevent a package dependency cycle
between cloud and install packages.