Give the monitor ssh sessions a valid (non-/dev/null) stdin
The ssh sessions used for monitoring were previously being passed the
equivalent of `/dev/null` for stdin. This `/dev/null` stdin was being
passed to `nc` which was causing it to exit immediately instead of the
desired blocking behavior. The obvious solution of using `nc -d` doesn't
work because it causes `nc` to spin internally (a `nc` bug?).