Expand the roachprod help text
Flesh out the help text for all of the roachprod commands.

Remove the use of SilenceUsage. This was preventing usage from being
shown when the positional arguments to a command were incorrect. Replace
the use of RunE with Run and wrap every run closure so that errors are
output but usage is not shown.

Fixes #64