ssh: permit multi-node ssh with iterm2 splits
This commit uses some AppleScript and iTerm2 magic to automatically pop
open a window with one split per node, and re-invoke roachprod ssh in
all of them. This allows you to have an ssh connection open to all of
the nodes in a roachprod cluster at once.

Then, you can use iTerm2's type-to-all-panes feature to interact with
all of the nodes at once (Shell->Broadcast Input->Broadcast Input to All
Panes in Current Tab, or Cmd-Ctrl-I).

Here's a gif of what it looks like:

https://gfycat.com/gifs/detail/LightheartedFavoriteFurseal