Add error message on 404 on `roachprod put local`
When `roachprod put local` is run no error is shown when trying to put a
file that does not exist because it is not copying the file, but trying
to make a symlink.

Instead of failing silently, now `roachprod put` will inform the user
that nothing was copied if the file doesn't exist.