Make the GCE project configurable
Add a `--gce-project` flag to the `create`, `destroy`, `list`, `sync`
and `gc` flags. Defaults to the `GCE_PROJECT` env var, or
`cockroach-ephemeral` if that env var is empty (or unset). Add a
`--gce-service-account` flag to the `create` command to allow
specification of a service account to use when creating nodes.

Fixes #136