Default to using scp for file transfers
Add a `--scp` flag which controls whether to use the builtin scp
implementation or the external `scp` tool. The default is to use the
external `scp` tool which allows for the use of compression and
hopefully avoids the mysterious `i/o timeout` errors. The downside to
using the external `scp` tool is the lack of a progress meter.

Fixes #142
Fixes #150