Add --treedist flag for put (defaults to true)
The treedist algorithm puts a file to multiple nodes in a cluster by
first copying the file to a single node, then using that node as a
source for "fanout" (2) additional copies. Then using the result
1+fanout nodes as sources for additional copies, etc.

The treedist algorithm can copy the 113MB cockroach binary to a 10 node
cluster 40% faster than the non-treedist approach from our corp
network. On a slower network the speedup should be even more dramatic.