Don't sync filtered cluster list to hosts file
Before this change, it was possible to see `unknown cluster` errors
when running multiple roachtest instances at once. The reason for
this was that any `roachprod list` command that included a filter
would only sync clusters matching that filter. This change removes
this bug by syncing the unfiltered cluster list instead.