Propagate exit status when running command on one node
When running a command on just one node, we can propagate its exit
status so that, e.g., `roachprod run CLUSTER:1 exit 7` exits with an
exit status of 7. Previously, that command would print

    Error: exit status 7

and then exit with error code 1. This made it difficult to
progamatically determine the exit status of the inner command.