Format multiple SSDs as single RAID 0 volume on AWS
This commit adjusts the AWS startup script to format and mount
multiple SSDs as a single volume using RAID 0. It follows the
instructions from
https://cloud.google.com/compute/docs/disks/local-ssd#formatmultiple.

This is important for AWS because many instance types come with
multiple instance store volumes by default. For instance, `m5d.4xlarge`
comes with 2 300GB local SSDs. The full list of defaults is in
https://docs.aws.amazon.com/AWSEC2/latest/UserGuide/InstanceStorage.html.

We'll probably want similar logic in GCE's startup script, although
that's not as important because GCE doesn't provide multiple SSDs
by default for any instance types.